# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SED Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Revenue report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [this_month, last_month, quarter, year]
            default: this_month
        - name: businessUnitId
          in: query
          schema:
            type: ["integer", "null"]
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month, business_unit, category]
            default: month
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /reports/analytics:
    get:
      operationId: getAnalytics
      tags: [reports]
      summary: Executive analytics (top/lowest performers)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [this_month, quarter, year]
            default: this_month
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsReport"
components:
  schemas:
...
    ChartDataPoint:
      type: object
      required: [label, revenue, expenses, profit]
      properties:
        label: { type: string }
        revenue: { type: number }
        expenses: { type: number }
        profit: { type: number }
...
    BusinessUnitStat:
      type: object
      required: [id, name, category, revenue, expenses, profit, growth]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        logoUrl: { type: ["string", "null"] }
        revenue: { type: number }
        expenses: { type: number }
        profit: { type: number }
        growth: { type: number }
...
    RevenueReport:
      type: object
      required: [totalRevenue, totalExpenses, netProfit, data]
      properties:
        totalRevenue: { type: number }
        totalExpenses: { type: number }
        netProfit: { type: number }
        data:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
    AnalyticsReport:
      type: object
      required: [topRevenue, lowestRevenue, highestExpense, highestProfit, customerGrowth, monthlyGrowth, yearlyGrowth]
      properties:
        topRevenue:
          type: array
          items:
            $ref: "#/components/schemas/BusinessUnitStat"
        lowestRevenue:
          type: array
          items:
            $ref: "#/components/schemas/BusinessUnitStat"
        highestExpense:
          type: array
          items:
            $ref: "#/components/schemas/BusinessUnitStat"
        highestProfit:
          type: array
          items:
            $ref: "#/components/schemas/BusinessUnitStat"
        customerGrowth: { type: number }
        monthlyGrowth: { type: number }
        yearlyGrowth: { type: number }